NOJIRA:  Disabling org.jasig.ssp.web.api.reference.ChallengeControllerHttpIntegrationTest.testControllerGet temporarily
Folks -- I'd like to disable this test for the time being so I/we can get to the
bottom of why the AbstractAuditableReferenceController.get(final @PathVariable UUID id)
method is not properly mapping to a request handler... for me.  I'm not working
anywhere near this code, so I think it's pretty likely a platform issue.  I'll
speak with John and/or others tomorrow about this item.

My apologies if I'm wrong and this is somehow something I didn't do properly.
Email to follow.