SSP-480. Present overwrite/reload option in UI when "new" student record already exists.
POST /api/1/person now returns a 409 if a user with the given school_id already exists.
Added special handling to the UI for that status code... the user can choose to just
go ahead and overwrite the server-side record, in which case the UI is tricked into
converting the POST to a PUT. Or the user can choose to load whatever is on the server
into the current form, losing her in-progress changes. To do this, I ended up just
reloading the entire view into edit mode. Tried introducing application events to
notify all the forms involved in the caseload assignment view that they needed to reload
their model, but that patch was becoming huge.

Still seem to be some issues with correctly saving all fields in the caseload assignment
form when in edit mode, but will handle these under separate tickets.