SSP-506. Avoid inadvertent username overwrites and incorrect person record updates.
Does't completely solve the problem (see SSP-521). But works around the biggest issue
by making sure that when you update record with a schoolId, you're actually updating
the record with that schoolId, not a record that *only* matched on username. Also,
whenever updating a person record, the already-persistent username is used instead
of the UI-calculated username. This should prevent issues like those that SPC
encountered where they elected the blind overwrite option when saving a caseload
assignment and the UI calculated username was updated to the person table, thereby
all of a sudden disabling that user's ability to log in.