NOJIRA. Suppress DEBUG logging from Jasper plugin.
Previously build output was terribly polluted with DEBUG logging output from the Jasper
plugin. Not sure exactly what was going on. All the obnoxious DEBUG logging is coming from
its usage of Jakarta Commons Digester. That thing using JCL, which should be fine b/c we
have jcl-over-slf4j as a Maven dependency. But for some reason the Jasper plugin
requires that dependency be specified again in its own configuration. It's seems to
be forking its own JVM... maybe that has something to do with it?