NOJIRA Increase scheduler pool size to match task count.
SPC has one extra task for synchronizing coaches in the background. This is partly
an attempt to address an as-of-yet undiagnosed permanent abort of
MessageServiceImpl.sendQueuedMessages() following database connection failure. Looking
at the Spring and java.util.concurrent code and testing in a dev env, a too-
small pool doesn't appear to affect rescheduling of that failed task, but we don't yet
have any better ideas given that we also can't see why the task would use anything
other than LoggingErrorHandler, which should just log and swallow exceptions. (As
opposed to PropagatingErrorHandler, which would log and raise, causing the task to
not reschedule).