SSP-542, SSP-570 Fix up db scripts.
Couldn't disentangle the SSP-542 and SSP-570 patches, so just kept both for the merge
down into 1.1.x despite SSP-542 being a feature add. The cherry pick ended up getting
too much of the 000012.xml script but too little (none) of masterChangeLog.xml. So this
attempts to fix that.
SSP-542 being a feature add