SSP-571 Include all assigned coaches in caseload report.
Previously if coaches were assigned to SSP-local person records but
those persons had no program status or did not have a student type
that satisfied the current filter, those coaches were excluded from
the caseload report entirely. But coaches reported as such by
uPortal were always included. This commit makes that handling
consistent.

Spent quite a bit of time trying to get this to work at the DAO
layer to avoid the separate call to the person DAO, but was just
too complicated to get it to work, largely b/c Hibernate makes
it very difficult to embed subselects in the FROM clause.