SSP-584 Fix hard-coded db name in Liquibase script 000014.xml.
To get this to work without duplicating the database name in
config, I introduced the ability to use property placeholders
in ssp-config.properties. So, for example, the default db
config now looks like:

  db_name=ssp
  db_url=jdbc:postgresql://127.0.0.1:5432/${db_name}

Also adds the ability to turn off that script by setting this in
ssp-config.properties:

  db_liquibase_set_mssql_snapshot_isolation=false

Felt that was needed b/c I've seen cases in dev where the command
never completes successfully (either when run from Liquibase or
the MSSQL management console). Seems to be related to trying to
run the command while other connections are still open to the db.
So seemed like it would be useful to let the app start up while
admins figured out how to apply the change out of band.