SSP-470 Merge and fix-up of the SSP-470 branch.
So now coaches as defined by uPortal are typically loaded in a
background job instead of as a side-effect of an API call.

The big problem with the SSP-470 branch was that it only worked
for SPC which happened to source its coach lists from a db table
instead of a uPortal API. When you try to query a uPortal API
(a REST API in particular) from a background job, it blows up
b/c there is no current HTTP request. This commit works around
that problem by just mocking up the HTTP request and response.
It depends on guest users having the appropriate uPortal
permissions to read the SSP_ROLES user attribute.