SSP-603 Add status code on external enrollment model.
Includes changes to Early Alert portlet to display the status code
after attempting to convert it to a more display-friendly name. The
name is configured with a new record in the config table that takes
a JSON string describing those code->name mappings. The default mappings
only have one mapping, which will translate a missing status code to
"Enrolled".

Removed the "Student Type" column from the EA portlet's roster view
and replaced it with the new "Status" column. Didn't see that
the student type column would ever be populated given the current APIs.