SSP-696 Use common ItemSelectorInitializer.
Works around issues with the ItemSelector widget incorrectly handling
selection of the last item. Previously copies of this code were
effectively embedded in SpecialServiceGroupsViewController.js and
EarlyAlertReferralsViewController.js. This commit refactors that
code into a shared class and adds usage of it to
EarlyAlertReferralsViewController (which is the actual fix for the
problem described in the Jira ticket).