SSP-749 Close EAs in same transaction as EA Response saves
Includes major refactorings to both front- and back-end code.
The back-end refactor was an attempt to move toward some of the proposals
here:
https://wiki.jasig.org/display/SSP/Technical+Design+Proposal:+Application+Layering

The front-end refactor had mostly to do with continued problems
with shared state in singleton models not being updated at the expected
times.

Note that involves a small API change for Early Alerts to include
the closing person name in GET responses. This allows the
EA detail screen to avoid a round trip to get that information.