SSP-805 Fix incorrect Early Alert Response targeting
Recent work moved the Early Alert Response creation button off of the
default Early Alert tool screen. By way of application events, though,
launch of the actual form was still handled by a member function
in the controller for the default list view
(EarlyAlertToolViewController). That handler function would look at
the list view's current SelectionModel, get the current Selection,
make sure that EA was loaded, and then load up the EA Response
form. The problem was that when the EA Response form was loaded the
default list view and controller were actually destroyed such that
when the user returned to the list view they actually got all
new view and controller objects under the covers. The application
event handler was still bound to the *old* view controller though.
So if you picked another EA then clicked on the button to create
a Response for it, the resulting event would be routed to the
*old* controller which would still be linked to the *old* view
which has no way of knowing that you've changed your EA selection.
So you'd get a EA Response form but it would be bound to
an unexpected EA.

The fix is just to dispense with application events and handle the
button click to launch the EA Response form right there in the
controller than owns that button (and which already knows the
correctly selected EA) (EarlyAlertDetailsViewController).