SSP-769, SSP-787 Add three student intake fields and their config.
Added 2 new configurations one for Weekly hours and one for registration
load.  Added fields for course_work_weekly_yours_name and
registration_load_name as well as anticipated_graduation_date_term_code.
UI complete for SSP

This is basically a cherry-pick and fix-up of
9f7f5f99e77805bea3b1caffad36cf9109905042, which needed to be moved
off of rel-1-2-patches while that branch was focused on the 1.2.0
release. Biggest fix-up issue was with removing patches that had to
do with the new photo feature in the Main tool.

Also aggregated and minified JS changes.