SSP-811 Avoid currentPerson race condition btwn Main and Action Plan
If you clicked the Action Plan tool before the Main tool had finished
loading its data, Action Plan would blow up with a 404. The latter
expected the 'currentPerson' singleton to be fully populated, which
Main normally does. But obviously that can't be depended on if it
is possible to interrupt loading to nav between tools.

The workaround was to modify all the Action Plan controllers to
use the 'personLite' singleton to get the selected user's ID.
That singleton is populated as a side-effect of selecting a user
in the search result panel. DisplayStrengthsViewController does
need access to the record represented by 'currentPerson', though,
so added code to perform that initialization as needed.

Also modified ProfilePersonViewController, which loads
currentPerson for the Main tool, to not wait for all its data to
return before populating models and views. This is an attempt to
reduce the chance of the same race popping up anywhere else we
might have a tool that assumes currentPerson is always fully
populated.

Conflicts:
	src/main/webapp/app-all.js
	src/main/webapp/app/controller/tool/profile/ProfilePersonViewController.js