SSP-893 Always dismiss program status change reason dialog
Dialog would remain onscreen if an error occurred in program status
change reason persistence failed. (Actually, the error that was
actually reported when we added logging here had nothing to do with
the save... seemed to have something to do with re-loading the
person afterwards. But for some reason this does seem to fix
the UX problem of the dialog remaining on screen. The status
change *is* persistent correctly.)