SSP-818 Refactor EAR Journal Entry comment template
Template is now plain text rather than HTML markup and is focused
on the Early Alert Response that generates instances of the
dereferenced template. Previously the template was focused on the
Early Alert to which the Response referred.

This commit reflects the final state of the fix in master and
rel-1-2, but actually cherry-picking those commits didn't make a
whole lot of sense, especially not from master b/c 000024.xml
ended up having another unrelated changeset added.

For related commits on other branches, see:

  c7159b26eae3ef1014776bf860664c4a6402f4d9
  4d2513230f63744b0e09a1e2f47e56a3ccb32143
  7fc13f1a96269501b8bb237ead716256f28b5b3c