SSP-651 Add aggr js to .gitignore
There's no good way to have the sencha command run during the
package phase so we can keep its input *and* output completely
confined to the target dir - the war plugin always beats us to
the punch and zips up the war after copying the webapp resources,
but before we have a chance to run. So... we just accept that
the aggr js will end up in source and add it to .gitignore.