SSP-1014 Namespace caseload projection column aliases.
This works around an issue with Hibernate+SqlServer where a
paginated query will break when the order by field references
a column alias rather than an explicit column name. This seems
to only be a problem on MSSQL because the Hibernate dialect
uses an OVER clause to contain the order by and TSQL
disallows alias references in such clauses.