SSP-973 Refactor objectStatus/active sync to fix 'add' behavior
Clicking 'Add' on the 'Electives Type' admin form was blowing up.
Still don't fully understand what was going on, but seems the
'record.set()' calls I was using to sync up these two fields are
dangerous when the object isn't yet bound to a store. Not clear
to me when exactly that happens... must be just after initialization
b/c set() is still being called before
AbstractReferenceAdminViewController.addRecord() ever does anything
to add the model instance it creates to a store explicitly. In
any event... the fix was basically to try to detect when we're in
initialization and skip the explicitly synchronization. As long
as Ext.js guarantees that fields are initialized in the same way
and order in which they're set during deserialization (which does
seem to be the case), this just happens to work fine...