SSP-1040 Override default Liquibase SQLServer type mappings
The default mapping was from 'date' to 'smalldatetime'. This doesn't
work when Hibernate maps such fields to/from a java.sql.Date. Plus
we don't really want any time component in these db fields. So for
2008 and greater we change the mapping to target the 'date'
physical type. 2005 and earlier don't have a 'date' type, so we just
want to use 'datetime' and let Hibernate deal with zeroing out the
time component.