SSP-1076 Migrate JVM-local timestamps to db_time_zone.
Slightly different processes for SQLServer and Postgres b/c
Postgres can do all the work in a single statement per table.

This applies to all the external and internal db table up
through 1.2, except those columns where we need to change the
actual data type.

We do not address any of the tables which are entirely new
to 2.0 b/c those have not yet been released to the wild
and, in reality, were introduced after the application
switched to using UTC by default. So there is no real point
in a migration for those tables.

Split up changeSets both by db and minor release. The
intention is to make it easy to back-port individual files
*and* avoid accidentally running a big changeset as part
of an upgrade that re-runs migrations against already
migrate timestamps (which would be a disaster).

Includes config files (scripts/timestamp-fields-*) used to
generate the changeSets via
timestamp_changeset_generator_*.groovy. Those groovy scripts
are intentionally quick and dirty.