SSP-908 Render printable map in MyGPS via popup
Uses the same template-configured markup as MAP printing in the Ext.js
app.

Still effectively disabled for most users right now b/c there are two
controller methods involved, both of which require permissions a
student isn't likely to have:

 1) PlanController#getCurrentForStudent() - Needs ROLE_PERSON_READ
 2) PlanController#print() - Needs ROLE_PERSON_MAP_READ

As is, though, the patch is no less secure than what we've currently
got and someone else will be making the necessary backend changes, so
I'm treating this commit as a complete unit of work as far as the
front-end is concerned.