SSP-1408 Inject missing EA coordinators into a now unpaginated combobox
So now the Campus create/edit screen will show *all* uP-designated
coaches and if the currently associated coach (if in edit mode) is not
in that list, it will make an attempt to find and inject that coach
into the list. That attempt should be nearly guaranteed to succeed b/c
we have FKs in the database which should ensure the corresponding person
record still exists, even if uPortal doesn't think that person is a
coach. Should that fail, though, the end user will see an "UNKNOWN
PERSON" entry selected in the EA Coordinator combobox.

De-paginating the combo box did not require a new store b/c we already
had an unpaginated Coaches.js. In fact, I decided to delete the store
that this screen was previously using (EarlyAlertCoordinators.js) b/c
it was proxying the same URL as Coaches.js, but had needlessly complicated
configuration. Only impact of this seemed to be a minor change to the
view definition to read the coach display name from a slightly
differently named model field.