SSP-1564 Do not reference view in controller destroy methods
View has already been destroyed by the time a controller is destroyed.
So trying to reference views from within controller destruction methods
was resulting in the JS equivalent of NPEs and the impression of
broken navigation b/c the current controller would just repeatedly
try and fail to destroy itself.