SSP-1205 Store and reload intake completed date after save
The model strips out a few fields before sending to the server due to issues it causes. The intake completed date was never put back into the model after that event. Thus, when saving and after refreshing the client-side model it wouldn't be there. This typically happend in a cancel event. Stored the date and put it back into the model after the save method calls the model's strip method. Then in the build section added a check to reformat the date and put it into the correct field in the case specified above.