NOJIRA Fix IE display issues student intake
This fixes most disappearing fields and other display issues in IE9/10 in the student intake. There is only one issue remaining and that is if you scroll all the way to the bottom of the main app rapidly and go back up rapidly the very topmost fields may be cut off. The fields will be immediately restored on a slow scroll, which most users will do naturally.