NOJIRA First steps toward Hib storage of OAuth2 Clients
Includes some abortive work on a SspOauth2ClientServiceImpl that
tried to perform some clever mapping between Clients and Persons.
But that didn't work if usernames differed at all, and we really
can't guarantee that those won't be fiddled with. So this commit
will probably need to undergo some surgery after getting the
Hibernate work done but before being merged to master.