SSP-1652 Reload authN-created Person in current Hib session
Think this fixes another occurrence of our chronic problem with SSP
not rendering on first-time login. I'm still not entirely clear on
what happened, but it was definitely a regression from OAuth2 changes.

The proximate cause was that when a first-time login resulted in lazy
creation of a Person record, that newly created Person record was
placed into SspUser's ThreadLocal, but subsequent accesses to that
Person in the same thread would fail because the Person was actually
created in a different Hibernate Session.
(PersonServiceImpl#createUserAccount() always spawns its own
transaction, which means it gets its own Hibernate Session, independent
of the Session opened in our view Filter.)

The solution was to issue another query from the surrounding
transaction to fine the newly created Person. That way the ThreadLocal
would get a Person associated with the Hibernate Session the current
thread actually knew about.

As to why exactly the OAuth2 changes caused this problem, I'm not sure.
My guess is that it has something to do with the dramatically reduced
scope of the SpringSecurity-protected URLs after adding OAuth2. We
used to protect everything, even images. Now we just protect the API.
So perhaps there was some sort of unintentionally beneficial race
condition or synchronization that occurred as a side-effect of firing
the entire SpringSecurity stack on every web request that went away
once we stopped caring about security on most requests. Would not be
shocked if there's still a race condition lurking that could continue to
cause problems.

This commit also includes two other semi-related changes. The
SspUser's ThreadLocal<Person> is marked transient to avoid errors when
Tomcat tries to serialize SspUser instances into session storage. And
SecurityServiceImpl#afterRequest() will not longer initiate a
transaction. While debugging I observed that that method was firing
even when just handling requests for static resources. So opening and
closing transactions was just a waste of resources.