SSP-1649 Fix Color edit cancel
The problem was that the color picker defined in AdminTreeMenus.js
was actually instantiated in that file rather than just being
*specified* with an xtype struct. So I externalized the class
definition into ColorPicker.js, gave it a custom xtype
(sspcolorpicker) and referenced it by that xtype from
AdminTreeMenus.js.

That the picker was being instantiated in AdminTreeMenus.js was
problematic because it meant the same picker was being used
over and over, even after it, along with its owning view, were
destroyed, e.g. by navigating to different admin UI. This was
preventing Ext.js from applying styles to it post-destruction,
which resulted in the stack trace originally reported in the
ticket.