SSP-1649 Skip sort after inline edit cancel
Might have been (and still be) a good reason for wanting to reapply
a sort after a cancel. But when the SSP-1649-reported error was
fixed such that the intended sorting actually did occur, I felt
it was poor UX because it would reapply the default sorting for
the current UI, not the most recently end-user-selected sorting.
So for example:

  1) The default sorting in the Elective UI is sortOrder, which
     works,
  2) Then you sort on name by clicking that column,
  3) Then you double click a row to edit, but then decide against
     it and press cancel,
  4) Your changes are abandoned as you'd expect, but the table
     goes back to being sorted by sortOrder, not name, which
     is not what you'd expect. You'd expect your last preferred
     sorting to be retained.

This same "preserve current sort" behavior is now present on
both "cancel" and "update" following an inline edit.