SSP-1695 Force set coachId and studentTypeId if null in caseload assignment
This handles the one-off condition where either or both the coachId and the studentTypeId are null in the model loaded in caseload assignment. The model normally doesn't allow setting of those two values if the original was null, and this overrides that behavior so that a user can save a student by overriding those originally null values.