SSP-1766 Key and constraint cleanup for external_student_test
Two sets of dbms-specific changesets for clearing and recreating
external_student_test keys/constraints. As noted in comments, this
change does try to take into account the fact that a different
approach had already been attempted in Linux/PG CI. But still likely
to have to deal with some manual cleanup in dev envs, at a minimum
with b/c of some md5 changes in the 'add discriminator column to
external_student_test' changeset.

Removes Java-coded Liquibase generators for deleting unnamed
keys/constraints. Ended up being able to do that from the XML.

This is a cherry-pick and squash of the following:

  75b565376d73d76c791bcab2929c6edaba65ff84
  b7b7643fed0a9476491025ee5b262bbda57c7d47
  8e98c5103ffbb94599506a854715c512ff480bd7
  f56dd087ec81a75ce85504036f3187d3616e5397
  272186d6279729b37ed890c508fc2437d1f367f1
  45b912b3514b12a525fb74ec678e5b488ad40805
  822b2ce69f92878d77345a2ac95fdd0fe7b13dec
  bb5719b4730af99fb1b7e9d905ec952584fcdb41

The only conflict was in masterChangeLog.xml since the patches came
from master where there had been additional changes there that we
did not want to include.

The core of the log message above was taken from
822b2ce69f92878d77345a2ac95fdd0fe7b13dec