SSP-1812 Skip constraint drop if no existing constraint (MSSQL)
This addresses a bit of a one-off situation where existing SQLServer
deployments may have skipped creation of the natural key in
external_student_test because they already had data in that table.
In this case the skipping of the pre-discriminator key has to be
permanent b/c that key will never work. Thus there is never an
existing key to drop and the Liquibase was failing to add the
post-discriminator key.