NOJIRA Add more room for residency in Details tab
The residency field label is too small and when even normal county names are ued the result is no space between the title and the value. Added a few more units to the label width to eliminate the problem and to be consistent with the rest of the formatting on the page.