SSP-1828 Default to Template objectStatus value in save dialog
In edit mode we want the "Active Template" checkbox to be checked or
not depending on current Template objectStatus. In create mode we
want the checkbox to not be checked, i.e. create an inactive Template,
e.g. to support a quasi-"draft" mode.

The create mode behavior just happened to work previously, but the edit
mode behavior did not. This was because the form was being reset
on "show", so it was mostly by accident that either case worked.

But when I moved the line that initializes the activeness checkbox
into the show event handler, I noticed that edit mode exhibited the
opposite behavior, i.e. always checked rather than always unchecked.

This was because the JSON deserialization was running against a
singleton Plan model. The latter extends from AbstractBase, and uses
populateFromGenericObject() on it to consume the JSON. Because the
target object had already been initialized once, though, the flags it
used to coordinate between 'active' and 'objectStatus' fields had
been set in such a way (statusFieldsInitialized=true) that when
'active' was then set again, and with a null or undefined value (which
will always be the case when reading an object back from the server),
it thought the change was coming from the UI, not the server, and would
overwrite the value of 'objectStaus' back to the default ('ACTIVE').