SSP-1830 Fix type mismatch for Student History Report
Code was upadated in the EarlyAlertTO which affected this report. Changed the EarlyAlertTO to include the reference transfer objects along side the updated code so this report could function and have easy access to reference TO's. Also fixed an issue with Journal Session Details running over the Journal comments field which caused text placed on top of eachother.