SSP-1839 Allow public->private template conversion without pub perm
Previously a user without MAP_PUBLIC_TEMPLATE_WRITE couldn't make any
changes to a public template. Now we'll allow such users to demote it
to private, which should be just fine, either because it's your own
template, or it was somebody else's in which case the edit creates a
copy.