SSP-1828, SSP-1760 Restore useful line from a2c7682
I reverted a2c7682 because I felt later work on SSP-1828 resulted in
simpler changes that accomplished the same thing w/r/t synchronization
of objectStatus and active fields in the client-side model. But this
line, which makes sure the defaulting of objectStatus is reflected
in the value of the active field, was a helpful change that SSP-1828
did not address. So this commit restores it.