SSP-1963 Remove dependencies on ununsed unminified CSS/JS files
Temporary change for the 2.2.0 release. There were a large number
of CSS and JS imports added in support of client-side column
filtering for the "dynamic labels" feature which did not make it
into 2.2.0. It's not obvious how we should add those files to the
minified/aggregates CSS and/or JS files we're already using. So
since we don't need those files anyway in 2.2.0, I'm just
commenting out the imports and the JS code that depends on them.